change minion RPC call to heartbeat consumer. Fixed unstable tests
