+++ inventory/src/main/java/org/opennms/horizon/inventory/component/MinionHeartbeatConsumer.java
@@ -30,18 +30,38 @@ package org.opennms.horizon.inventory.component;
 
 import java.util.Map;
 import java.util.Optional;
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
 
-import io.grpc.Context;
+import javax.annotation.PreDestroy;
+
+import org.apache.kafka.clients.producer.ProducerRecord;
+import org.apache.kafka.common.header.internals.RecordHeader;
+import org.opennms.cloud.grpc.minion.RpcRequestProto;
+import org.opennms.cloud.grpc.minion.RpcResponseProto;
+import org.opennms.horizon.grpc.echo.contract.EchoRequest;
+import org.opennms.horizon.grpc.echo.contract.EchoResponse;
 import org.opennms.horizon.grpc.heartbeat.contract.HeartbeatMessage;
 import org.opennms.horizon.inventory.exception.InventoryRuntimeException;
 import org.opennms.horizon.inventory.service.MonitoringSystemService;
 import org.opennms.horizon.shared.constants.GrpcConstants;
+import org.opennms.taskset.contract.MonitorResponse;
+import org.opennms.taskset.contract.MonitorType;
+import org.opennms.taskset.contract.TaskResult;
+import org.opennms.taskset.contract.TaskSetResults;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.messaging.handler.annotation.Headers;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.stereotype.Component;
 
+import com.google.common.base.Strings;
+import com.google.protobuf.Any;
+import com.google.protobuf.InvalidProtocolBufferException;
+
+import io.grpc.Context;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -50,8 +70,17 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @PropertySource("classpath:application.yml")
 public class MinionHeartbeatConsumer {
+
+    private static final String DEFAULT_TASK_RESULTS_TOPIC = "task-set.results";
+    private static final int DEFAULT_MESSAGE_SIZE = 1024;
+    private static final long ECHO_TIMEOUT = 30_000;
+    private final MinionRpcClient rpcClient;
+    private final KafkaTemplate<String, byte[]> kafkaTemplate;
+
+    @Value("${kafka.topics.results:" + DEFAULT_TASK_RESULTS_TOPIC + "}")
+    private String kafkaTopic;
     private final MonitoringSystemService service;
-    private final MinionRpcManager minionRpc;
+
     @KafkaListener(topics = "${kafka.topics.minion-heartbeat}", concurrency = "1")
     public void receiveMessage(@Payload byte[] data, @Headers Map<String, Object> headers) {
         try {
@@ -60,12 +89,72 @@ public class MinionHeartbeatConsumer {
                 .orElseThrow(()->new InventoryRuntimeException("Missing tenant id"));
             log.info("Received heartbeat message for minion with tenant id: {}; id: {}; location: {}", tenantId, message.getIdentity().getSystemId(), message.getIdentity().getLocation());
             Context.current().withValue(GrpcConstants.TENANT_ID_CONTEXT_KEY, tenantId).run(()->
-            {
-                Optional<Long> newCreatedId = service.addMonitoringSystemFromHeartbeat(message, tenantId);
-                newCreatedId.ifPresent(minionRpc::addSystem);
-            });
+                service.addMonitoringSystemFromHeartbeat(message, tenantId));
+            CompletableFuture.runAsync(() -> runRpcMonitor(message.getIdentity().getSystemId(), message.getIdentity().getLocation(), tenantId));
         } catch (Exception e) {
             log.error("Error while processing heartbeat message: ", e);
         }
+    }
+
+    @PreDestroy
+    public void shutDown() {
+        if(rpcClient != null) {
+            rpcClient.shutdown();
+        }
+    }
+
+    private void runRpcMonitor(String systemId, String location, String tenantId) {
+        log.info("Sending RPC request for minion {} with location {}", systemId, location);
+        EchoRequest echoRequest = EchoRequest.newBuilder()
+            .setTime(System.nanoTime())
+            .setMessage(Strings.repeat("*", DEFAULT_MESSAGE_SIZE))
+            .build();
+        RpcRequestProto request = RpcRequestProto.newBuilder()
+            .setLocation(location)
+            .setSystemId(systemId)
+            .setModuleId("Echo")
+            .setExpirationTime(System.currentTimeMillis() + ECHO_TIMEOUT)
+            .setRpcId(UUID.randomUUID().toString())
+            .setPayload(Any.pack(echoRequest))
+            .build();
+        rpcClient.sendRpcRequest(tenantId, request).thenApply(RpcResponseProto::getPayload)
+            .thenApply(payload -> {
+                try {
+                    return payload.unpack(EchoResponse.class);
+                } catch (InvalidProtocolBufferException e) {
+                    throw new RuntimeException(e);
+                }
+            })
+            .whenComplete((echoResponse, error) -> {
+                    if (error != null) {
+                        log.error("Unable to complete echo request", error);
+                        return;
+                    }
+                    long responseTime = (System.nanoTime() - echoResponse.getTime()) / 1000000;
+                    publishResult(systemId, location, tenantId, responseTime);
+                    log.info("Response time for minion {} is {} msecs", systemId, responseTime);
+                }
+            );
+    }
+
+    private void publishResult(String systemId, String location, String tenantId, long responseTime) {
+        MonitorResponse monitorResponse = MonitorResponse.newBuilder()
+            .setStatus("UP")
+            .setResponseTimeMs(responseTime)
+            .setMonitorType(MonitorType.ECHO)
+            .setIpAddress(systemId) //for minion only
+            .build();
+        TaskResult result = TaskResult.newBuilder()
+            .setId("monitor-"+systemId)
+            .setSystemId(systemId)
+            .setLocation(location)
+            .setMonitorResponse(monitorResponse)
+            .build();
+        TaskSetResults results = TaskSetResults.newBuilder()
+            .addResults(result)
+            .build();
+        ProducerRecord<String, byte[]> producerRecord = new ProducerRecord<>(kafkaTopic, results.toByteArray());
+        producerRecord.headers().add(new RecordHeader(GrpcConstants.TENANT_ID_KEY, tenantId.getBytes()));
+        kafkaTemplate.send(producerRecord);
     }
 }
