+++ inventory/src/test/java/org/opennms/horizon/inventory/service/taskset/response/DetectorResponseServiceIntTest.java
@@ -1,12 +1,19 @@
 package org.opennms.horizon.inventory.service.taskset.response;
 
-import io.grpc.Context;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.io.IOException;
+import java.net.InetAddress;
+import java.net.UnknownHostException;
+import java.time.LocalDateTime;
+import java.util.List;
+
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.opennms.horizon.inventory.SpringContextTestInitializer;
-import org.opennms.horizon.inventory.dto.MonitoredServiceDTO;
 import org.opennms.horizon.inventory.grpc.GrpcTestBase;
 import org.opennms.horizon.inventory.grpc.taskset.TestTaskSetGrpcService;
 import org.opennms.horizon.inventory.model.IpInterface;
@@ -19,7 +26,6 @@ import org.opennms.horizon.inventory.repository.MonitoredServiceRepository;
 import org.opennms.horizon.inventory.repository.MonitoredServiceTypeRepository;
 import org.opennms.horizon.inventory.repository.MonitoringLocationRepository;
 import org.opennms.horizon.inventory.repository.NodeRepository;
-import org.opennms.horizon.inventory.service.MonitoredServiceService;
 import org.opennms.horizon.shared.constants.GrpcConstants;
 import org.opennms.horizon.shared.utils.InetAddressUtils;
 import org.opennms.taskset.contract.DetectorResponse;
@@ -30,14 +36,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ContextConfiguration;
 
-import java.io.IOException;
-import java.net.InetAddress;
-import java.net.UnknownHostException;
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import io.grpc.Context;
 
 
 @SpringBootTest
@@ -66,15 +65,17 @@ class DetectorResponseServiceIntTest extends GrpcTestBase {
     @Autowired
     private MonitoredServiceRepository monitoredServiceRepository;
 
-    @Autowired
-    private MonitoredServiceService monitoredServiceService;
-
     private static TestTaskSetGrpcService testGrpcService;
 
     @BeforeAll
     public static void setup() throws IOException {
         testGrpcService = new TestTaskSetGrpcService();
         server = startMockServer(TaskSetServiceGrpc.SERVICE_NAME, testGrpcService);
+    }
+
+    @BeforeEach
+    void beforeTest() {
+        testGrpcService.reset();
     }
 
     @AfterEach
@@ -87,8 +88,6 @@ class DetectorResponseServiceIntTest extends GrpcTestBase {
             nodeRepository.deleteAll();
             monitoringLocationRepository.deleteAll();
         });
-
-        testGrpcService.reset();
     }
 
     @AfterAll
@@ -125,7 +124,7 @@ class DetectorResponseServiceIntTest extends GrpcTestBase {
         assertEquals(response.getMonitorType().name(), monitoredServiceType.getServiceName());
         assertEquals(TEST_TENANT_ID, monitoredServiceType.getTenantId());
 
-        List<MonitoredService> monitoredServices = monitoredServiceRepository.findAll();;
+        List<MonitoredService> monitoredServices = monitoredServiceRepository.findAll();
         assertEquals(1, monitoredServices.size());
 
         MonitoredService monitoredService = monitoredServices.get(0);
@@ -141,7 +140,7 @@ class DetectorResponseServiceIntTest extends GrpcTestBase {
     }
 
     @Test
-    void testAcceptMultipleSameIpAddress() throws UnknownHostException {
+    void testAcceptMultipleSameIpAddress() {
         Context.current().withValue(GrpcConstants.TENANT_ID_CONTEXT_KEY, TEST_TENANT_ID).run(()->
         {
             try {
@@ -165,7 +164,7 @@ class DetectorResponseServiceIntTest extends GrpcTestBase {
     }
 
     @Test
-    void testAcceptNotDetected() throws UnknownHostException {
+    void testAcceptNotDetected() {
         Context.current().withValue(GrpcConstants.TENANT_ID_CONTEXT_KEY, TEST_TENANT_ID).run(()->
         {
             try {
@@ -221,7 +220,7 @@ class DetectorResponseServiceIntTest extends GrpcTestBase {
             List<String> monitoredServiceNames =
                 monitoredServiceTypes.stream()
                     .map(MonitoredServiceType::getServiceName)
-                    .collect(Collectors.toList());
+                    .toList();
 
             assertEquals(monitorTypes.length, monitoredServiceNames.size());
